+++ src/Subjects/GroupedSubject.php
@@ -22,10 +22,9 @@ namespace TechDivision\Import\Product\Grouped\Subjects;
 
 use TechDivision\Import\Utils\RegistryKeys;
 use TechDivision\Import\Product\Subjects\AbstractProductSubject;
-use TechDivision\Import\Product\Link\Exceptions\MapSkuToEntityIdException;
 
 /**
- * A subject implementation the process to import grouped products.
+ * A subject implementation to process the import of grouped products.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
  * @copyright 2019 TechDivision GmbH <info@techdivision.com>
@@ -35,13 +34,6 @@ use TechDivision\Import\Product\Link\Exceptions\MapSkuToEntityIdException;
  */
 class GroupedSubject extends AbstractProductSubject
 {
-
-    /**
-     * The mapping for the SKUs to the created entity IDs.
-     *
-     * @var array
-     */
-    protected $skuEntityIdMapping = array();
 
     /**
      * Intializes the previously loaded global data for exactly one variants.
@@ -62,31 +54,7 @@ class GroupedSubject extends AbstractProductSubject
         // load the status of the actual import process
         $status = $registryProcessor->getAttribute(RegistryKeys::STATUS);
 
-        // load the attribute set we've prepared intially
+        // load the SKU => entity ID mapping
         $this->skuEntityIdMapping = $status[RegistryKeys::SKU_ENTITY_ID_MAPPING];
-    }
-
-    /**
-     * Return the entity ID for the passed SKU.
-     *
-     * @param string $sku The SKU to return the entity ID for
-     *
-     * @return integer The mapped entity ID
-     * @throws \TechDivision\Import\Product\Link\Exceptions\MapSkuToEntityIdException Is thrown if the SKU is not mapped yet
-     */
-    public function mapSkuToEntityId($sku)
-    {
-
-        // query weather or not the SKU has been mapped
-        if (isset($this->skuEntityIdMapping[$sku])) {
-            return $this->skuEntityIdMapping[$sku];
-        }
-
-        // throw an exception if the SKU has not been mapped yet
-        throw new MapSkuToEntityIdException(
-            $this->appendExceptionSuffix(
-                sprintf('Found not mapped entity ID for SKU %s', $sku)
-            )
-        );
     }
 }